+++ src/main/java/com/onnoff/onnoff/auth/feignClient/client/KakaoOauth2Client.java
@@ -1,12 +1,13 @@
-package com.onnoff.onnoff.auth.client;
+package com.onnoff.onnoff.auth.feignClient.client;
 
 
-import com.onnoff.onnoff.auth.client.dto.KakaoOauth2DTO;
+import com.onnoff.onnoff.auth.feignClient.config.FeignConfig;
+import com.onnoff.onnoff.auth.feignClient.dto.KakaoOauth2DTO;
 import feign.Headers;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
 
-@FeignClient(name = "kakao-oauth2-client", url = "https://kauth.kakao.com")
+@FeignClient(name = "kakao-oauth2-client", url = "https://kauth.kakao.com", configuration = FeignConfig.class)
 public interface KakaoOauth2Client {
     @Headers("Content-Type: application/x-www-form-urlencoded")
     @PostMapping("/oauth/token")
